pyramidPix2pix/requirements.txt
beautifulsoup4==4.12.3
dominate==2.9.1
numpy==1.25.0
opencv_python==4.7.0.72
Pillow==10.3.0
Requests==2.31.0
scipy==1.12.0
scikit-image>=0.18.0
torch==2.0.1
torchvision==0.15.2
visdom==0.2.4
wandb==0.16.5
xmltodict
kornia